filter

+++ zic_mod_filter1.h
@@ -15,23 +15,22 @@ protected:
         FILTER_MODE_COUNT,
     };
 
-    int16_t buf0 = 0.0;
-    int16_t buf1 = 0.0;
-    int16_t buf2 = 0.0;
-    int16_t buf3 = 0.0;
+    // buf can be int16
+    float buf0 = 0.0;
+    float buf1 = 0.0;
+    float buf2 = 0.0;
+    float buf3 = 0.0;
 
     float cutoffCalc = 0.2;
     // NOTE maybe set feedback manually, just to try
-    // or at least try mult 2 or 3
-    // also cutoff cannot be 1.0 else div by zero, maybe use mult instead
-    uint16_t feedback;
+    // feedback can be uint16
+    float feedback;
     void calculateVar()
     {
         cutoffCalc = cutoff;
         // cutoffCalc = 1.0f - cutoff * cutoff;
         feedback = resonance + resonance / (1.0 - cutoffCalc);
-        // Serial.printf("cutoff %.2f calc %.2f res %.2f feedback %.2f\n", cutoff, cutoffCalc, resonance, feedback);
-        Serial.printf("cutoff %.2f calc %.2f res %.2f feedback %d\n", cutoff, cutoffCalc, resonance, feedback);
+        Serial.printf("cutoff %.2f calc %.2f res %.2f feedback %.2f\n", cutoff, cutoffCalc, resonance, feedback);
     }
 
 public:
